Fix apparent issues when using boost with nonstandard compiler
Even though Clang aims to be ABI compatible with GCC, apparently Boost libraries compiled with homebrew are somehow not compatible with GCC (at least 11 and 12 from Hombrew). The solution is to only use the Boost headers and to compile the gzip and zlib parts of boost as part of the IMSRG code. This step guarantees that when we change compilers we can simply recompile the boost lib and have it be guaranteed to be ABI compatible with the rest of the code/libraries.

What a nightmare.